feat: enhance stack configurability
