+++ cdktf.json
@@ -1,6 +1,6 @@
 {
   "language": "typescript",
-  "app": "npx ts-node main.ts",
+  "app": "npx ts-node test.ts",
   "projectId": "116578e3-9930-4a16-8805-b23da1f7d156",
   "sendCrashReports": "false",
   "terraformProviders": [],
